feat: Add API integration for generating reports
This commit adds API integration to the ReportsPage component. It sends a GET request to the backend admin server to fetch the report data based on the selected time period and report type. The generateTimePeriod function is introduced to generate the appropriate time period query parameter for the API request. The dummyReports object is removed as it is no longer needed. The axios import is updated to use the custom axios instance from the axios.js file. The fetched report data is stored in the reportData state variable.